角色绑定菜单、绑定权限功能完善

+++ pyod-service/src/main/java/top/yigege/service/impl/RoleServiceImpl.java
@@ -5,6 +5,8 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.transaction.annotation.Transactional;
 import top.yigege.constant.BusinessFlagEnum;
+import top.yigege.dao.MenuMapper;
+import top.yigege.model.Permission;
 import top.yigege.model.Role;
 import top.yigege.dao.RoleMapper;
 import top.yigege.model.User;
@@ -13,11 +15,16 @@ import top.yigege.service.IRoleService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.stereotype.Service;
 import top.yigege.util.PageUtil;
+import top.yigege.vo.LayuiTreeBean;
 import top.yigege.vo.PageBean;
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+
+import static top.yigege.constant.PyodConstant.Common.PARENT_MENU_FLAG;
 
 /**
  * <p>
@@ -36,6 +43,9 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
     @Resource
     RoleMapper roleMapper;
 
+    @Resource
+    MenuMapper menuMapper;
+
     @Override
     public Role queryRoleInfo(Integer roleId) {
        return roleMapper.queryRoleInfo(roleId);
@@ -45,19 +55,26 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
     @Override
     public void bindRolePermission(Integer roleId, List<Integer> permissionIds) {
         //1.先解绑所有角色
-        roleMapper.deleteRolePermissionRecord(roleId);
-
+        if (null != roleId) {
+         roleMapper.deleteRolePermissionRecord(roleId);
+        }
         //2. 绑定现有角色
-        roleMapper.addRolePermissionRecord(roleId, permissionIds);
+        if (!permissionIds.isEmpty()) {
+            roleMapper.addRolePermissionRecord(roleId, permissionIds);
+        }
     }
 
+    @Transactional
     @Override
     public void bindRoleMenu(Integer roleId, List<Integer> menuIds) {
         //1.先解绑所有菜单
-        roleMapper.deleteRoleMenuRecord(roleId);
-
+        if (null != roleId) {
+          roleMapper.deleteRoleMenuRecord(roleId);
+        }
         //2. 绑定现有菜单
-        roleMapper.addRoleMenuRecord(roleId, menuIds);
+        if (!menuIds.isEmpty()) {
+            roleMapper.addRoleMenuRecord(roleId, menuIds);
+        }
     }
 
     @Override
@@ -91,10 +108,95 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
         }
         saveOrUpdate(role);
 
-        bindRoleMenu(role.getRoleId(),menuIds);
+        if (!menuIds.isEmpty()) {
+            bindRoleMenu(role.getRoleId(),menuIds);
+        }
 
-        bindRolePermission(role.getRoleId(),permissionIds);
+        if (!permissionIds.isEmpty()) {
+            bindRolePermission(role.getRoleId(),permissionIds);
+        }
 
         return queryRoleInfo(role.getRoleId());
+    }
+
+    @Override
+    public List<LayuiTreeBean> queryCheckedMenusByRoleId(Integer roleId) {
+        List<LayuiTreeBean> totalMenuList = roleMapper.queryTreeMenusByRoleId(roleId);
+
+        List<LayuiTreeBean> menuTree = new ArrayList<>();
+        for (LayuiTreeBean menu : totalMenuList) {
+            if (PARENT_MENU_FLAG == menu.getPid()) {
+                menuTree.add(findChildren(menu, totalMenuList));
+            }
+        }
+
+        if (!menuTree.isEmpty()) {
+            //对该菜单按照sort排序
+            Collections.sort(menuTree);
+        }
+
+
+        return menuTree;
+    }
+
+    @Override
+    public List<LayuiTreeBean> queryMenusByRoleId(Integer roleId) {
+        List<LayuiTreeBean> totalMenuList = menuMapper.queryTreeMenu();
+
+        List<LayuiTreeBean> totalCheckedMenuList = roleMapper.queryTreeMenusByRoleId(roleId);
+
+        //设置当前已选中
+        for (LayuiTreeBean currentCheckMenu : totalCheckedMenuList) {
+            for (LayuiTreeBean item : totalMenuList) {
+                if (currentCheckMenu.getId() == item.getId()) {
+                    item.setChecked(true);
+                    break;
+                }
+            }
+        }
+
+        //建树
+        List<LayuiTreeBean> menuTree = new ArrayList<>();
+        for (LayuiTreeBean menu : totalMenuList) {
+            if (PARENT_MENU_FLAG == menu.getPid()) {
+                menuTree.add(findChildren(menu, totalMenuList));
+            }
+        }
+
+        if (!menuTree.isEmpty()) {
+            //对该菜单按照sort排序
+            Collections.sort(menuTree);
+        }
+
+        return menuTree;
+    }
+
+    @Override
+    public List<Permission> queryAllPermissionByRoleId(Integer roleId) {
+        return roleMapper.queryAllPermissionByRoleId(roleId);
+    }
+
+
+    /**
+     * 递归查找子节点
+     * @param menu
+     * @param menuList
+     * @return
+     */
+    private LayuiTreeBean findChildren(LayuiTreeBean menu, List<LayuiTreeBean> menuList) {
+
+        for (LayuiTreeBean menuItem : menuList) {
+            //是否存在menu的子节点
+            if (menu.getId() == menuItem.getPid()) {
+                menu.getChildren().add(findChildren(menuItem, menuList));
+            }
+        }
+
+        if (!menu.getChildren().isEmpty()) {
+            //对该菜单的所有子节点按照sort排序
+            Collections.sort(menu.getChildren());
+        }
+
+        return menu;
     }
 }
